Misc updates.

+++ src/map/map.c
@@ -2569,11 +2569,11 @@ int map_addinstancemap(const char *name, int id)
 	map[dst_m].npc_num = 0;
 
 	// Reallocate cells
-	num_cell = map[dst_m].xs * map[dst_m].ys;
+	num_cell = (size_t)map[dst_m].xs * map[dst_m].ys;
 	CREATE(map[dst_m].cell, struct mapcell, num_cell);
 	memcpy(map[dst_m].cell, map[src_m].cell, num_cell * sizeof(struct mapcell));
 
-	size = map[dst_m].bxs * map[dst_m].bys * sizeof(struct block_list *);
+	size = (size_t)(map[dst_m].bxs * map[dst_m].bys) * sizeof(struct block_list *);
 	map[dst_m].block = (struct block_list **)aCalloc(1,size);
 	map[dst_m].block_mob = (struct block_list **)aCalloc(1,size);
 
@@ -3587,9 +3587,9 @@ int map_readgat (struct map_data *m)
 	if (gat == NULL)
 		return 0;
 
-	m->xs = *(int32*)(gat+6);
-	m->ys = *(int32*)(gat+10);
-	num_cells = m->xs * m->ys;
+	m->xs = *(int32 *)(gat + 6);
+	m->ys = *(int32 *)(gat + 10);
+	num_cells = (size_t)m->xs * m->ys;
 	CREATE(m->cell, struct mapcell, num_cells);
 
 	water_height = map_waterheight(m->name);
@@ -3697,7 +3697,7 @@ int map_readallmaps (void)
 		map[i].bxs = (map[i].xs + BLOCK_SIZE - 1) / BLOCK_SIZE;
 		map[i].bys = (map[i].ys + BLOCK_SIZE - 1) / BLOCK_SIZE;
 
-		size = map[i].bxs * map[i].bys * sizeof(struct block_list *);
+		size = (size_t)(map[i].bxs * map[i].bys) * sizeof(struct block_list *);
 		map[i].block = (struct block_list**)aCalloc(size, 1);
 		map[i].block_mob = (struct block_list**)aCalloc(size, 1);
 	}
