+++ src/tool/mapcache.c
@@ -117,7 +117,7 @@ void cache_map(char *name, struct map_data *m)
 	len = (unsigned long)m->xs*(unsigned long)m->ys*2;
 	write_buf = (unsigned char *)aMalloc(len);
 	// Compress the cells and get the compressed length
-	encode_zip(write_buf, &len, m->cells, m->xs*m->ys);
+	encode_zip(write_buf, &len, m->cells, (uint64)m->xs*m->ys);
 
 	// Fill the map header
 	if (strlen(name) > MAP_NAME_LENGTH) // It does not hurt to warn that there are maps with name longer than allowed.
